Moved user related components used by containers to components folder
