+++ src/containers/UserList/UserFilter.js
@@ -5,7 +5,7 @@ import i18next from 'i18next';
 import { Checkbox } from 'material-ui';
 import DropDown from 'd2-ui/lib/form-fields/DropDown.component';
 import { updateFilter, getList } from '../../actions';
-import OrganisationUnitInput from './OrganisationUnitInput';
+import OrganisationUnitInput from '../../components/OrganisationUnitInput';
 import SearchFilter from '../../components/SearchFilter';
 import {
     INACTIVE_MONTHS,
